+++ python/tests/sanity_checks/test_sanity_checks.py
@@ -1,6 +1,6 @@
 import os
 import pandas as pd
-from sanity_checks.sanity_checks import check_pct_values
+from sanity_checks.sanity_check import check_pct_values
 
 
 THIS_DIR = os.path.dirname(os.path.abspath(__file__))
